main/vessel/requirements.txt
numpy>=1.20
scipy>=1.7
scikit-image>=0.18
Pillow>=8.0
matplotlib>=3.3
